Bugfix: reoganise configuration settings
