+++ trae_agent/utils/anthropic_client.py
@@ -4,7 +4,6 @@
 """Anthropic API client wrapper with tool integration."""
 
 import json
-import os
 import random
 import time
 from typing import override
@@ -23,14 +22,6 @@ class AnthropicClient(BaseLLMClient):
 
     def __init__(self, model_parameters: ModelParameters):
         super().__init__(model_parameters)
-
-        if self.api_key == "":
-            self.api_key: str = os.getenv("ANTHROPIC_API_KEY", "")
-
-        if self.api_key == "":
-            raise ValueError(
-                "Anthropic API key not provided. Set ANTHROPIC_API_KEY in environment variables or config file."
-            )
 
         self.client: anthropic.Anthropic = anthropic.Anthropic(
             api_key=self.api_key, base_url=self.base_url
