+++ trae_agent/utils/openai_client.py
@@ -4,7 +4,6 @@
 """OpenAI API client wrapper with tool integration."""
 
 import json
-import os
 import random
 import time
 from typing import override
@@ -28,19 +27,6 @@ class OpenAIClient(BaseLLMClient):
 
     def __init__(self, model_parameters: ModelParameters):
         super().__init__(model_parameters)
-
-        if self.api_key == "":
-            self.api_key: str = os.getenv("OPENAI_API_KEY", "")
-
-        if self.api_key == "":
-            raise ValueError(
-                "OpenAI API key not provided. Set OPENAI_API_KEY in environment variables or config file."
-            )
-
-        if "OPENAI_BASE_URL" in os.environ:
-            # If OPENAI_BASE_URL is set, which means the user wants to use a specific openai compatible api provider,
-            # we should use the base url from the environment variable
-            self.base_url = os.environ["OPENAI_BASE_URL"]
 
         self.client: openai.OpenAI = openai.OpenAI(api_key=self.api_key, base_url=self.base_url)
         self.message_history: ResponseInputParam = []
