+++ trae_agent/utils/openrouter_client.py
@@ -37,14 +37,6 @@ class OpenRouterClient(BaseLLMClient):
     def __init__(self, model_parameters: ModelParameters):
         super().__init__(model_parameters)
 
-        if self.api_key == "":
-            self.api_key: str = os.getenv("OPENROUTER_API_KEY", "")
-
-        if self.api_key == "":
-            raise ValueError(
-                "OpenRouter API key not provided. Set OPENROUTER_API_KEY in environment variables or config file."
-            )
-
         # Use OpenAI SDK with OpenRouter's base URL
         self.client: openai.OpenAI = openai.OpenAI(api_key=self.api_key, base_url=self.base_url)
         self.message_history: list[ChatCompletionMessageParam] = []
